tapnode/tapnodemock: hoist node-side mocks out of tapgarden
MockChainBridge, MockWalletAnchor, MockKeyRing, and
GenMockGroupVerifier implement the tapnode interfaces but lived in
tapgarden/mock.go. They were consumed across tapgarden, tapfreighter,
and universe/supplyverifier tests; the tapgarden home was an
accident of original authorship.

Move them into tapnode/tapnodemock as ChainBridge, WalletAnchor,
KeyRing, and GenGroupVerifier (the package-name-prefix stutter is
dropped on the way through). The NewGenesisTx / FundGenesisTx
helpers that lived alongside MockWalletAnchor move with it.
tapnode/tapnodemock is a subpackage of tapnode so production code
cannot transitively depend on the mocks; tests opt in by importing
the subpackage explicitly.
